# rest-api/public/OpenAPISpec.json
{
	"components": {
...
		"schemas": {
			"IntId": {
				"type": "integer",
				"format": "int32",
				"description": "An integer that serves as a record id. Used for primary keys and foreign keys.",
				"minimum": 1
			},
...
			"Genre": {
				"description": "Genre",
				"properties": {
					"GenreId": {
						"$ref": "#/components/schemas/IntId"
					},
					"Name": {
						"type": "string"
					}
				},
				"required": [
					"GenreId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ValidateErrorJSON": {
				"description": "A validation error",
				"properties": {
					"message": {
						"type": "string",
						"enum": [
							"Validation failed"
						],
						"nullable": false
					},
					"details": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					}
				},
				"required": [
					"message",
					"details"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NotFoundError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Customer": {
				"description": "Customer",
				"properties": {
					"CustomerId": {
						"$ref": "#/components/schemas/IntId"
					},
					"FirstName": {
						"type": "string"
					},
					"LastName": {
						"type": "string"
					},
					"Company": {
						"type": "string"
					},
					"Address": {
						"type": "string"
					},
					"City": {
						"type": "string"
					},
					"State": {
						"type": "string"
					},
					"Country": {
						"type": "string"
					},
					"PostalCode": {
						"type": "string"
					},
					"Phone": {
						"type": "string"
					},
					"Fax": {
						"type": "string"
					},
					"Email": {
						"type": "string"
					},
					"SupportRepId": {
						"$ref": "#/components/schemas/IntId"
					}
				},
				"required": [
					"CustomerId",
					"FirstName",
					"LastName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CustomerInsertInput": {
				"description": "Customer Insert Input",
				"properties": {
					"FirstName": {
						"type": "string"
					},
					"LastName": {
						"type": "string"
					},
					"Company": {
						"type": "string"
					},
					"Address": {
						"type": "string"
					},
					"City": {
						"type": "string"
					},
					"State": {
						"type": "string"
					},
					"Country": {
						"type": "string"
					},
					"PostalCode": {
						"type": "string"
					},
					"Phone": {
						"type": "string"
					},
					"Fax": {
						"type": "string"
					},
					"Email": {
						"type": "string"
					},
					"SupportRepId": {
						"$ref": "#/components/schemas/IntId"
					}
				},
				"required": [
					"FirstName",
					"LastName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CustomerUpdateInput": {
				"description": "Customer Update Input",
				"properties": {
					"FirstName": {
						"type": "string"
					},
					"LastName": {
						"type": "string"
					},
					"Company": {
						"type": "string"
					},
					"Address": {
						"type": "string"
					},
					"City": {
						"type": "string"
					},
					"State": {
						"type": "string"
					},
					"Country": {
						"type": "string"
					},
					"PostalCode": {
						"type": "string"
					},
					"Phone": {
						"type": "string"
					},
					"Fax": {
						"type": "string"
					},
					"Email": {
						"type": "string"
					},
					"SupportRepId": {
						"$ref": "#/components/schemas/IntId"
					}
				},
				"required": [
					"FirstName",
					"LastName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Employee": {
				"description": "Employee",
				"properties": {
					"EmployeeId": {
						"$ref": "#/components/schemas/IntId"
					},
					"LastName": {
						"type": "string"
					},
					"FirstName": {
						"type": "string"
					},
					"Title": {
						"type": "string"
					},
					"ReportsTo": {
						"$ref": "#/components/schemas/IntId"
					},
					"BirthDate": {
						"type": "string",
						"format": "date-time"
					},
					"HireDate": {
						"type": "string",
						"format": "date-time"
					},
					"Address": {
						"type": "string"
					},
					"City": {
						"type": "string"
					},
					"State": {
						"type": "string"
					},
					"Country": {
						"type": "string"
					},
					"PostalCode": {
						"type": "string"
					},
					"Phone": {
						"type": "string"
					},
					"Fax": {
						"type": "string"
					},
					"Email": {
						"type": "string"
					}
				},
				"required": [
					"EmployeeId",
					"LastName",
					"FirstName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EmployeeInsertInput": {
				"description": "Employee Insert Input",
				"properties": {
					"Title": {
						"type": "string"
					},
					"FirstName": {
						"type": "string"
					},
					"LastName": {
						"type": "string"
					},
					"Address": {
						"type": "string"
					},
					"City": {
						"type": "string"
					},
					"State": {
						"type": "string"
					},
					"Country": {
						"type": "string"
					},
					"PostalCode": {
						"type": "string"
					},
					"Phone": {
						"type": "string"
					},
					"Fax": {
						"type": "string"
					},
					"Email": {
						"type": "string"
					},
					"ReportsTo": {
						"$ref": "#/components/schemas/IntId"
					},
					"BirthDate": {
						"type": "string",
						"format": "date-time"
					},
					"HireDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"FirstName",
					"LastName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EmployeeUpdateInput": {
				"description": "Employee Update Input",
				"properties": {
					"Title": {
						"type": "string"
					},
					"FirstName": {
						"type": "string"
					},
					"LastName": {
						"type": "string"
					},
					"Address": {
						"type": "string"
					},
					"City": {
						"type": "string"
					},
					"State": {
						"type": "string"
					},
					"Country": {
						"type": "string"
					},
					"PostalCode": {
						"type": "string"
					},
					"Phone": {
						"type": "string"
					},
					"Fax": {
						"type": "string"
					},
					"Email": {
						"type": "string"
					},
					"ReportsTo": {
						"$ref": "#/components/schemas/IntId"
					},
					"BirthDate": {
						"type": "string",
						"format": "date-time"
					},
					"HireDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"FirstName",
					"LastName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GenreInsertInput": {
				"description": "Genre Insert Input",
				"properties": {
					"Name": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"GenreUpdateInput": {
				"description": "Genre Update Input",
				"properties": {
					"Name": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "rest-api",
		"version": "1.0.0",
		"description": "Sample REST API built on the Chinook database, for Hasura demo",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "benoit ranque"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/customers": {
			"get": {
				"operationId": "GetCustomers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Customer"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get a set of resources. Can be paginated",
				"tags": [
					"Customer"
				],
				"security": [],
				"parameters": [
					{
						"description": "how many records should be returned",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 20,
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					},
					{
						"description": "how many records to offset by",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				]
			},
			"post": {
				"operationId": "InsertCustomer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Customer"
								}
							}
						}
					}
				},
				"description": "Create a new instance of the resource. Returns the created resource.",
				"tags": [
					"Customer"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Request Body",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CustomerInsertInput",
								"description": "Request Body"
							}
						}
					}
				}
			}
		},
		"/customers/{CustomerId}": {
			"get": {
				"operationId": "GetCustomer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Customer"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					},
					"422": {
						"description": "Validation failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					}
				},
				"description": "Get a single resource using the resource id.",
				"tags": [
					"Customer"
				],
				"security": [],
				"parameters": [
					{
						"description": "resource id",
						"in": "path",
						"name": "CustomerId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/IntId"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateCustomer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Customer"
								}
							}
						}
					}
				},
				"description": "Update the resource using the resource id. Returns the updated resource.",
				"tags": [
					"Customer"
				],
				"security": [],
				"parameters": [
					{
						"description": "resource id",
						"in": "path",
						"name": "CustomerId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/IntId"
						}
					}
				],
				"requestBody": {
					"description": "Request Body",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CustomerUpdateInput",
								"description": "Request Body"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCustomer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Customer"
								}
							}
						}
					}
				},
				"description": "Delete the resource using the resource id. Returns the deleted resource.",
				"tags": [
					"Customer"
				],
				"security": [],
				"parameters": [
					{
						"description": "resource id",
						"in": "path",
						"name": "CustomerId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/IntId"
						}
					}
				]
			}
		},
		"/employees": {
			"get": {
				"operationId": "GetEmployees",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Employee"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get a set of resources. Can be paginated",
				"tags": [
					"Employee"
				],
				"security": [],
				"parameters": [
					{
						"description": "how many records should be returned",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 20,
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					},
					{
						"description": "how many records to offset by",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				]
			},
			"post": {
				"operationId": "InsertEmployee",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Employee"
								}
							}
						}
					}
				},
				"description": "Create a new instance of the resource. Returns the created resource.",
				"tags": [
					"Employee"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EmployeeInsertInput"
							}
						}
					}
				}
			}
		},
		"/employees/{EmployeeId}": {
			"get": {
				"operationId": "GetEmployee",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Employee"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					},
					"422": {
						"description": "Validation failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					}
				},
				"description": "Get a single resource using the resource id.",
				"tags": [
					"Employee"
				],
				"security": [],
				"parameters": [
					{
						"description": "resource id",
						"in": "path",
						"name": "EmployeeId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/IntId"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateEmployee",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Employee"
								}
							}
						}
					}
				},
				"description": "Update the resource using the resource id. Returns the updated resource.",
				"tags": [
					"Employee"
				],
				"security": [],
				"parameters": [
					{
						"description": "resource id",
						"in": "path",
						"name": "EmployeeId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/IntId"
						}
					}
				],
				"requestBody": {
					"description": "Request Body",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EmployeeUpdateInput",
								"description": "Request Body"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteEmployee",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Employee"
								}
							}
						}
					}
				},
				"description": "Delete the resource using the resource id. Returns the deleted resource.",
				"tags": [
					"Employee"
				],
				"security": [],
				"parameters": [
					{
						"description": "resource id",
						"in": "path",
						"name": "EmployeeId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/IntId"
						}
					}
				]
			}
		},
		"/genres": {
			"get": {
				"operationId": "GetGenres",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Genre"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get a set of resources. Can be paginated",
				"tags": [
					"Genre"
				],
				"security": [],
				"parameters": [
					{
						"description": "how many records should be returned",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 20,
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					},
					{
						"description": "how many records to offset by",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					}
				]
			},
			"post": {
				"operationId": "InsertGenre",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Genre"
								}
							}
						}
					}
				},
				"description": "Create a new instance of the resource. Returns the created resource.",
				"tags": [
					"Genre"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Request Body",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GenreInsertInput",
								"description": "Request Body"
							}
						}
					}
				}
			}
		},
		"/genres/{GenreId}": {
			"get": {
				"operationId": "GetGenre",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Genre"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					},
					"422": {
						"description": "Validation failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					}
				},
				"description": "Get a single resource using the resource id.",
				"tags": [
					"Genre"
				],
				"security": [],
				"parameters": [
					{
						"description": "resource id",
						"in": "path",
						"name": "GenreId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/IntId"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateGenre",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Genre"
								}
							}
						}
					}
				},
				"description": "Update the resource using the resource id. Returns the updated resource.",
				"tags": [
					"Genre"
				],
				"security": [],
				"parameters": [
					{
						"description": "resource id",
						"in": "path",
						"name": "GenreId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/IntId"
						}
					}
				],
				"requestBody": {
					"description": "Request Body",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GenreUpdateInput",
								"description": "Request Body"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteGenre",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Genre"
								}
							}
						}
					}
				},
				"description": "Delete the resource using the resource id. Returns the deleted resource.",
				"tags": [
					"Genre"
				],
				"security": [],
				"parameters": [
					{
						"description": "resource id",
						"in": "path",
						"name": "GenreId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/IntId"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}